+++ mono/src/main/java/woowa/team4/bff/menu/category/dto/create/MenuCategoryCreateRequest.java
@@ -1,13 +1,11 @@
-package woowa.team4.bff.menu.category.dto;
+package woowa.team4.bff.menu.category.dto.create;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-@AllArgsConstructor
 public class MenuCategoryCreateRequest {
 
     @NotBlank
@@ -15,7 +13,7 @@ public class MenuCategoryCreateRequest {
     @NotBlank
     private String description;
 
-    public MenuCategoryDto toMenuCategoryDto(final String restaurantUuid) {
-        return new MenuCategoryDto(restaurantUuid, name, description);
+    public MenuCategoryCreateDto toDto(final String restaurantUuid) {
+        return new MenuCategoryCreateDto(restaurantUuid, name, description);
     }
 }
